+++ Startup/macros.h
@@ -42,11 +42,11 @@ assert(!"only 1 target compiler can be specified")
 
 #if HANDMADE_LOCAL_BUILD
 
- //#define HANDMADE_DEBUG_FPS
- //#define HANDMADE_DEBUG_CLOCKCYCLES
- //#define HANDMADE_DEBUG_AUDIO
+#define HANDMADE_DEBUG_FPS
+//#define HANDMADE_DEBUG_CLOCKCYCLES
+//#define HANDMADE_DEBUG_AUDIO
 #define HANDMADE_LIVE_LOOP_EDITING
-#define HANDMADE_DEBUG_LIVE_LOOP_EDITING
+//#define HANDMADE_DEBUG_LIVE_LOOP_EDITING
 //#define HANDMADE_DEBUG_TILE_POS
 //#define HANDMADE_WALK_THROUGH_WALLS
